+++ tests/spec/from_dict_test.py
@@ -4,6 +4,7 @@ import mock
 import simplejson as json
 from six.moves.urllib import parse as urlparse
 
+from bravado_core.model import MODEL_MARKER
 from bravado_core.spec import Spec
 
 
@@ -14,7 +15,7 @@ def test_definitions_not_present(minimal_swagger_dict):
 
 
 @mock.patch('jsonref.JsonRef')
-@mock.patch('bravado_core.spec.replace_jsonref_proxies')
+@mock.patch('bravado_core.spec.post_process_spec')
 @mock.patch.object(Spec, 'build')
 def test_origin_uri_gets_passed_to_jsonref(mock_build, mock_prox, mock_ref,
                                            minimal_swagger_dict):
@@ -49,5 +50,55 @@ def test_relative_ref_spec():
 
     # TODO: Make the behavior of x-model consitent for both the specs. Currently
     # it gets populated only for the former. Hence, removing that key from dict
-    delete_key_from_dict(expected_dict, 'x-model')
+    delete_key_from_dict(expected_dict, MODEL_MARKER)
     assert expected_dict == json.loads(json.dumps(resultant_spec.spec_dict))
+
+
+def test_ref_to_external_path_with_ref_to_local_model():
+    # Test that an an external ref to a path (in swagger.json) which contains
+    # a local ref to a model (in pet.json) works as expected:
+    # - model type for Pet is created
+    # - de-reffed spec_dict contains 'x-model' annotations
+    my_dir = os.path.abspath(os.path.dirname(__file__))
+
+    swagger_json_path = os.path.join(
+        my_dir,
+        '../../test-data/2.0/x-model/swagger.json')
+
+    with open(swagger_json_path) as f:
+        swagger_json_content = json.loads(f.read())
+
+    swagger_json_url = urlparse.urljoin('file:', swagger_json_path)
+    spec = Spec.from_dict(swagger_json_content, swagger_json_url)
+
+    assert spec.definitions['Pet']
+    assert spec.spec_dict['paths']['/pet']['get']['responses']['200'][
+        'schema'][MODEL_MARKER] == 'Pet'
+
+
+def test_spec_with_dereffed_and_tagged_models_works(minimal_swagger_dict):
+    # In cases where the Swagger spec being ingested has already been de-reffed
+    # and had models tagged with 'x-model', we still need to be able to
+    # detect them and make them available as model types. For example, a spec
+    # ingested via http from pyramid_swagger contains de-reffed models.
+    pet_path_spec = {
+        'get': {
+            'responses': {
+                '200': {
+                    'description': 'Returns a Pet',
+                    'schema': {
+                        MODEL_MARKER: 'Pet',
+                        'type': 'object',
+                        'properties': {
+                            'name': {
+                                'type': 'string'
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+    minimal_swagger_dict['paths']['/pet'] = pet_path_spec
+    spec = Spec.from_dict(minimal_swagger_dict)
+    assert spec.definitions['Pet']